apps/analytics-backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.31.1
PyJWT==2.9.0
pydantic==2.9.2
